字典转模型修改

+++ Runtime/runtime-字典转模型/runtime-字典转模型/Model/PeopleModel.h
@@ -12,7 +12,7 @@
 
 @property (nonatomic, copy) NSString *name;
 
-@property (nonatomic, assign) NSInteger hight;
+@property (nonatomic, assign) float hight;
 
 @property (nonatomic, strong) NSNumber *wight;
